Created AdminActivityController and AdminUserController and organized all http methods
