+++ appsettings.json
@@ -1,7 +1,7 @@
 {
     "ConnectionStrings": {
-        //"DefaultConnection": "Server=sql.dennym.de,1433; Database=dbSBIFS; User Id=sa; Password=Baum876543;TrustServerCertificate=True;"
-        "DefaultConnection": "Server=Localhost,1433; Database=dbSBIFS; User Id=sa; Password=1234;TrustServerCertificate=True;"
+        "DefaultConnection": "Server=sql.dennym.de,1433; Database=dbSBIFS; User Id=sa; Password=Baum876543;TrustServerCertificate=True;"
+        //"DefaultConnection": "Server=Localhost,1433; Database=dbSBIFS; User Id=sa; Password=1234;TrustServerCertificate=True;"
     },
     "Logging": {
         "LogLevel": {
